+++ src/stdlib.rs
@@ -367,6 +367,11 @@ fn is_empty(vm: &mut VM, value: RantValue) -> RantStdResult {
   Ok(())
 }
 
+fn is_any(vm: &mut VM, value: RantValue) -> RantStdResult {
+  vm.cur_frame_mut().write_value(RantValue::Boolean(!value.is_empty()));
+  Ok(())
+}
+
 fn is_bool(vm: &mut VM, value: RantValue) -> RantStdResult {
   vm.cur_frame_mut().write_value(RantValue::Boolean(value.get_type() == RantValueType::Boolean));
   Ok(())
@@ -1414,7 +1419,7 @@ pub(crate) fn load_stdlib(context: &mut Rant)
     // Verification functions
     is_string as "is-string", is_integer as "is-integer", is_float as "is-float", 
     is_number as "is-number", is_bool as "is-bool", is_empty as "is-empty", is_nan as "is-nan",
-    is_between as "is-between",
+    is_between as "is-between", is_any as "is-any",
 
     // Math functions
     add, sub, mul, div, mul_add as "mul-add", mod_ as "mod", neg, recip, is_odd as "is-odd", is_even as "is-even", is_factor as "is-factor",